refactor(driver): improve driver extract and update logic
